Revamped Weapons Systems and added new Weapons
Reworked a lot of Weapon_Template to get it up to date.

Set it up to easily create weapons by just creating child classes and
editing a couple values.

Added 'Spread' Which allows a weapon to possibly not be 100% accurate
all of the time. (AKA it allows weapons to not be a lazer bean and shoot
in the same spot therefore making weapons a bit more realistic? maybe?
Idk its cool though)

Added two new weapons testing out the new weapon_template class. these
weapons are shotgun and machinegun (pretty general i know..) and it
worked beautifully. YAY